Refactor styles: remove unused imports and obsolete ProfileItemDelegate code and fixed missing KeyBindings
